Made this App work with multiple Python versions
It's now possible to have different Python versions be installed and
used in parallel.
Added a download manager, where you can download the different python
versions as well as additional Python modules. This allows for a greater
customisation of what gets installed and how much storage space is used.

Also added a main menu and some code documentation.

+++ app/src/main/java/com/apython/python/pythonhost/Util.java
@@ -1,13 +1,17 @@
 package com.apython.python.pythonhost;
 
 import android.content.Context;
+import android.support.annotation.NonNull;
 import android.util.Log;
 
+import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
-import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.DefaultHttpClient;
+import org.apache.http.params.BasicHttpParams;
+import org.apache.http.params.HttpConnectionParams;
+import org.apache.http.params.HttpParams;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -20,16 +24,30 @@ import java.io.PrintWriter;
 import java.lang.reflect.Array;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.security.DigestInputStream;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
+import java.util.Scanner;
 
 /*
- * This class provides a bunch of general useful utility unctions.
+ * This class provides a bunch of general useful utility functions.
  *
  * Created by Sebastian on 05.07.2015.
  */
 
 public class Util {
 
+    /**
+     * Ensures that a file or directory can be accessed from other apps.
+     * More specific, ensures that other apps have rad and execute,
+     * but not write permission.
+     *
+     * @param file The file or directory, that should be made accessible.
+     * @param recursive If {@code true}, makes all files and subdirectories of the given
+     *                  directory accessible, if the given FileObject is a directory.
+     * @return {@code true} on success, {@code false} otherwise.
+     */
     public static boolean makeFileAccessible(File file, boolean recursive) {
         if (!file.exists()) {
             try {
@@ -38,8 +56,7 @@ public class Util {
                     return false;
                 }
             } catch (IOException e) {
-                Log.w(MainActivity.TAG, "Could not make file '" + file.getAbsolutePath() + "' accessible: Could not create file.");
-                e.printStackTrace();
+                Log.w(MainActivity.TAG, "Could not make file '" + file.getAbsolutePath() + "' accessible: Could not create file.", e);
                 return false;
             }
         }
@@ -60,13 +77,21 @@ public class Util {
                 return true;
             }
         } catch (IOException e) {
-            e.printStackTrace();
+            Log.e(MainActivity.TAG, "Failed to make '" + file.getAbsolutePath() + "' accessible via chmod!", e);
         } catch (InterruptedException ie) {
-            ie.printStackTrace();
+            Log.e(MainActivity.TAG, "Failed to make '" + file.getAbsolutePath() + "' accessible via chmod: Chmod was interrupted!", ie);
         }
         return false;
     }
 
+    /**
+     * Checks if the given directory and all it's content
+     * are accessible from other apps. More specific, checks
+     * that all apps have the right to read and execute.
+     *
+     * @param dir The directory that should be checked.
+     * @return {@code true}, if the directory and all it's contents are accessible.
+     */
     public static ArrayList<File> checkDirContentsAccessibility(File dir) {
         ArrayList<File> inaccessibleFiles = new ArrayList<>();
         ProcessBuilder processBuilder = new ProcessBuilder("ls", "-l").directory(dir);
@@ -92,38 +117,80 @@ public class Util {
         return inaccessibleFiles;
     }
 
-    public static HttpResponse connectToUrl(String url) {
+    /**
+     * Connects to an url and returns the {@code HttpResponse} on success.
+     *
+     * @param url The url to connect to.
+     * @param timeout The maximum amount of time to wait for the connection to be established.
+     * @return A {@code HttpResponse} on success and {@code null} on failure.
+     */
+    public static HttpResponse connectToUrl(String url, int timeout) {
         HttpResponse response = null;
         try {
-            HttpClient client = new DefaultHttpClient();
+            final HttpParams httpParams = new BasicHttpParams();
+            HttpConnectionParams.setConnectionTimeout(httpParams, timeout * 1000);
+            HttpClient client = new DefaultHttpClient(httpParams);
             HttpGet request = new HttpGet();
             request.setURI(new URI(url));
             response = client.execute(request);
         } catch (URISyntaxException uriE) {
-            uriE.printStackTrace();
-        } catch (ClientProtocolException cpE) {
-            cpE.printStackTrace();
-        } catch (IOException ioE) {
-            ioE.printStackTrace();
+            Log.e(MainActivity.TAG, "Failed to connect to '" + url + "': Invalid url!", uriE);
+        } catch (IOException e) {
+            Log.e(MainActivity.TAG, "Failed to connect to '" + url + "'!", e);
         }
         return response;
     }
 
-    public static boolean installRawResource(File destination, InputStream resourceLocation, PackageManager.ProgressHandler progressHandler) {
+    /**
+     * See {@link #installFromInputStream(File, InputStream, long, ProgressHandler)}.
+     *
+     * @param destination The destination where the {@code InputStream} should be saved.
+     * @param resourceLocation The {@code InputStream} to read from.
+     * @param progressHandler An optional progress handler (can be {@code null}).
+     * @return {@code true}, if the installation succeeded, {@code false} otherwise.
+     */
+    public static boolean installFromInputStream(File destination, InputStream resourceLocation, ProgressHandler progressHandler) {
+        long size;
+        try {
+            size = resourceLocation.available();
+        } catch (IOException e) {
+            Log.e(MainActivity.TAG, "Failed to install resource to " + destination.getAbsolutePath() + ": Failed to get the estimated length of the input.", e);
+            return false;
+        }
+        return installFromInputStream(destination, resourceLocation, size, progressHandler);
+    }
+
+    /**
+     * Installs the contents of the given {@code InputStream} and saves it to the specified
+     * destination. Uses the {@code ProgressHandler} to report progress.
+     *
+     * @param destination The destination where the {@code InputStream} should be saved.
+     * @param resourceLocation The {@code InputStream} to read from.
+     * @param inputLength The total amount of bytes that can be read from the {@code InputStream}.
+     * @param progressHandler An optional progress handler (can be {@code null}).
+     * @return {@code true}, if the installation succeeded, {@code false} otherwise.
+     */
+    public static boolean installFromInputStream(File destination, InputStream resourceLocation, long inputLength, ProgressHandler progressHandler) {
         byte[] buffer = new byte[1024];
         int count;
         try {
-            int total = resourceLocation.available();
             int totalCount = 0;
             float nextUpdate = 0;
-            float onePercent = (float) total / 100;
+            float onePercent = inputLength / 100.0f;
+            if (!destination.getParentFile().exists() && !destination.getParentFile().mkdirs()) {
+                Log.w(MainActivity.TAG, "Failed to install resource to " + destination.getAbsolutePath() + ": Could not make the directories!");
+                return false;
+            }
+            if (progressHandler != null) {
+                progressHandler.setProgress(0);
+            }
             FileOutputStream outputFile = new FileOutputStream(destination);
             while ((count = resourceLocation.read(buffer)) != -1) {
                 outputFile.write(buffer, 0, count);
                 if (progressHandler != null) {
                     totalCount += count;
                     if (totalCount >= nextUpdate) {
-                        progressHandler.setProgress((float) totalCount / total);
+                        progressHandler.setProgress((float) totalCount / inputLength);
                         nextUpdate += onePercent;
                     }
                 }
@@ -134,15 +201,96 @@ public class Util {
             outputFile.close();
             Util.makeFileAccessible(destination, false);
         } catch (IOException e) {
-            Log.e(MainActivity.TAG, "Failed to extract the raw resource to " + destination.getAbsolutePath() + "!");
+            Log.e(MainActivity.TAG, "Failed to install resource to " + destination.getAbsolutePath() + "!");
             e.printStackTrace();
             return false;
         }
         return true;
     }
 
+    /**
+     * Downloads a file from a specified url and saves the contents in the
+     * given destination. The content will be checked against the given
+     * md5CheckSum. Progress is reported via the optional {@code ProgressHandler}.
+     *
+     * @param url The url from which to download from.
+     * @param destination The file to save the contents of the data retrieved from the url to.
+     * @param md5CheckSum The md5 check sum the downloaded data must have in order to be valid.
+     * @param progressHandler An optional progress handler (can be {@code null}).
+     * @return {@code true} if the data was successfully downloaded, verified and made accessible
+     * to other apps.
+     */
+    public static boolean downloadFile(String url, File destination, String md5CheckSum, ProgressHandler progressHandler) {
+        HttpResponse response = Util.connectToUrl(url, 30);
+        if (response == null) {
+            Log.e(MainActivity.TAG, "Failed to connect to '" + url + "'!");
+            return false;
+        }
+        HttpEntity entity = response.getEntity();
+        MessageDigest hashAlgorithm;
+        try {
+            hashAlgorithm = MessageDigest.getInstance("MD5");
+        } catch (NoSuchAlgorithmException e) {
+            Log.e(MainActivity.TAG, "Failed to download from '" + url + "': Md5 Hash is not supported on this device!", e);
+            return false;
+        }
+        try {
+            DigestInputStream inputStream = new DigestInputStream(entity.getContent(), hashAlgorithm);
+            entity.getContentLength();
+            if (!Util.installFromInputStream(destination, inputStream, entity.getContentLength(), progressHandler)) {
+                return false;
+            }
+            if (!hashToHex(hashAlgorithm.digest()).equals(md5CheckSum)) {
+                Log.w(MainActivity.TAG, "Md5 checksum of downloaded File invalid!");
+                if (!destination.delete()) {
+                    Log.e(MainActivity.TAG, "Failed to delete downloaded file '" + destination.getAbsolutePath() + "'.");
+                }
+                return false;
+            }
+        } catch (IOException e) {
+            Log.e(MainActivity.TAG, "Failed to download '" + destination.getName() + "'!", e);
+            if (destination.exists()) {
+                if (!destination.delete()) {
+                    Log.w(MainActivity.TAG, "Failed to clean up downloaded file '" + destination.getAbsolutePath() + "'.");
+                }
+            }
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Converts a hash into a hex string.
+     *
+     * @param hash The hash to convert
+     * @return A hexadecimal representation of the hash.
+     */
+    public static String hashToHex(byte[] hash) {
+        StringBuilder hexString = new StringBuilder();
+        for (byte hashByte : hash) {
+            String hex = Integer.toHexString(0xff & hashByte);
+            if (hex.length() == 1) hexString.append('0');
+            hexString.append(hex);
+        }
+        return hexString.toString();
+    }
+
+    /**
+     * Add two arrays together.
+     *
+     * @param first The first array.
+     * @param second The second array.
+     * @param <T> The type of array.
+     * @return A new array that contains all the elements of {@code first} and {@code second}.
+     */
     public static <T> T[] mergeArrays(T[] first, T[] second) {
-        if (second == null) { return first; }
+        if (second == null) {
+            if (first != null) {
+                return first;
+            } else {
+                throw new IllegalArgumentException("No array given to mergeArrays.");
+            }
+        }
         if (first == null) { return second; }
 
         int aLen = first.length;
@@ -156,7 +304,14 @@ public class Util {
         return result;
     }
 
-    public static void parsePipOutput(Context context, PackageManager.ProgressHandler progressHandler, String text) {
+    /**
+     * Parse output given by pip.
+     *
+     * @param context The current context.
+     * @param progressHandler A progressHandler to report the output to.
+     * @param text A line of output from pip.
+     */
+    public static void parsePipOutput(Context context, ProgressHandler progressHandler, String text) {
         text = text.trim().replace("[?25h", "");
         if (text.startsWith("[K")) {
             String[] parts = text.split(" ");
@@ -183,5 +338,68 @@ public class Util {
             progressHandler.setText(context.getString(R.string.module_installation_building));
         }
         Log.d(MainActivity.TAG, text);
+    }
+
+    /**
+     * Delete a directory and all it's content.
+     *
+     * @param directory The directory to delete.
+     * @return {@code true} on success, {@code false} otherwise.
+     */
+    public static boolean deleteDirectory(File directory) {
+        File[] files = directory.listFiles();
+        if (files != null) {
+            for (File file : files) {
+                if (file.isDirectory()) {
+                    if (!deleteDirectory(file)) {
+                        return false;
+                    }
+                } else {
+                    if (!file.delete()) {
+                        Log.w(MainActivity.TAG, "Failed to delete file '" + file.getAbsolutePath() + "'.");
+                        return false;
+                    }
+                }
+            }
+        }
+        if (directory.exists() && !directory.delete()) {
+            Log.w(MainActivity.TAG, "Failed to delete directory '" + directory.getAbsolutePath() + "'.");
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Converts the content of a given {@code InputStream} to a string.
+     *
+     * @param input An {@code InputStream} to read from.
+     * @return The content of the given {@code InputStream} as a {@code String}
+     */
+    public static String convertStreamToString(InputStream input) {
+        if (input == null) { return ""; }
+        Scanner s = new Scanner(input).useDelimiter("\\A");
+        return s.hasNext() ? s.next() : "";
+    }
+
+    /**
+     * Returns the first and the second part of a python version.
+     *
+     * @param version A python Version
+     * @return The first and the second part of a python version.
+     */
+    public static String getMajorMinorVersionPart(@NonNull String version) {
+        String[] versionParts = version.split("\\.");
+        return versionParts[0] + "." + versionParts[1];
+    }
+
+    /**
+     * Converts a library File to the library name.
+     *
+     * @param libFile the library file
+     * @return The name of the library
+     */
+    public static String getLibraryName(File libFile) {
+        String libFileName = libFile.getName();
+        return libFileName.substring(3, libFileName.lastIndexOf('.'));
     }
 }
